source/requirements.txt
pandas==1.3.5
pytest==6.2.5
PyHamcrest==2.0.3
turfpy==0.0.7
geojson==2.5.0
geopandas==0.10.2
moto==2.3.0